requirements.txt
beautifulsoup4==4.11.1
Flask==2.2.2
matplotlib==3.6.0
Pillow==9.2.0
requests==2.28.1
spotipy==2.20.0
